+++ fsictools.py
@@ -10,7 +10,7 @@ for dependencies additional to those of `fsic`.
 from fsic import __version__
 
 import re
-from typing import List
+from typing import Dict, List
 
 from fsic import BaseModel, Symbol
 
@@ -20,6 +20,22 @@ def symbols_to_dataframe(symbols: List[Symbol]) -> 'pandas.DataFrame':
     from pandas import DataFrame
 
     return DataFrame([s._asdict() for s in symbols])
+
+def symbols_to_sympy(symbols: List[Symbol]) -> Dict['sympy.Symbol', 'sympy.Eq']:
+    """Convert the system of equations into a dictionary of `SymPy` objects. **Requires `SymPy`**."""
+    import sympy
+
+    system = {}
+
+    equations = [s.equation for s in symbols if s.equation is not None]
+    for e in equations:
+        e = e.replace('[t]', '')
+        e = re.sub(r'\[t[-]([0-9]+)\]', r'_\1', e)
+
+        lhs, rhs = map(sympy.sympify, e.split('=', maxsplit=1))
+        system[lhs] = sympy.Eq(lhs, rhs)
+
+    return system
 
 def model_to_dataframe(model: BaseModel) -> 'pandas.DataFrame':
     """Return the values and solution information from the model as a `pandas` DataFrame. **Requires `pandas`**."""